Make SKOS export for vocabularies available to non-admin users.
This fixes #1167

The baseURI is also now configurable and set by default to conform
to the standards team's recommendation, though it can still be
overridden if necessary.
