+++ conf/application.conf
@@ -308,6 +308,15 @@ ehri {
             holders: []
         }
     }
+
+    # SKOS/RDF export config. This sets the default baseURI for
+    # SKOS-format exported vocabularies and (potentially) other
+    # types of data.
+    lod {
+        CvocVocabulary {
+            baseUri: "http://data.ehri-project.eu/vocabularies/"
+        }
+    }
 }
 
 # system-specific overrides and extensions
