+++ modules/portal/app/views/vocabulary/itemDetails.scala.html
@@ -7,8 +7,14 @@
             @Markdown(desc)
         }
     } {
-        @views.html.common.typeLabelWithWatchButton(item, watched.contains(item.id), introNotice())
-        @views.html.common.childItemSidebar(item, EntityType.Concept)
+        @views.html.common.typeLabelWithWatchButton(item, watched.contains(item.id), views.html.vocabulary.introNotice())
+        @views.html.common.childItemSidebar(item, defines.EntityType.Concept)
+        @views.html.common.exportItem(
+            Map(
+                "ttl" -> controllers.portal.routes.Vocabularies.exportSkos(item.id, format = Some("TTL")),
+                "rdf_xml" -> controllers.portal.routes.Vocabularies.exportSkos(item.id, format = Some("RDF/XML")),
+            )
+        )
         @views.html.common.latestAction(item, controllers.portal.routes.Portal.itemHistory(item.id, modal = true))
     }
 }
