+++ modules/portal/conf/portal.routes
@@ -56,6 +56,7 @@ GET         /groups/:id                           @controllers.portal.Groups.bro
 GET         /vocabularies                         @controllers.portal.Vocabularies.searchAll(params: services.search.SearchParams ?= services.search.SearchParams.empty, paging: utils.PageParams ?= utils.PageParams.empty)
 GET         /vocabularies/:id                     @controllers.portal.Vocabularies.browse(id: String, params: services.search.SearchParams ?= services.search.SearchParams.empty, paging: utils.PageParams ?= utils.PageParams.empty)
 GET         /vocabularies/:id/search              @controllers.portal.Vocabularies.search(id: String, params: services.search.SearchParams ?= services.search.SearchParams.empty, paging: utils.PageParams ?= utils.PageParams.empty, inline: Boolean ?= false)
+GET         /vocabularies/:id/export              @controllers.portal.Vocabularies.exportSkos(id: String, format: Option[String] ?= None, baseUri: Option[String] ?= None)
 GET         /sets                                 @controllers.portal.AuthoritativeSets.searchAll(params: services.search.SearchParams ?= services.search.SearchParams.empty, paging: utils.PageParams ?= utils.PageParams.empty)
 GET         /sets/:id                             @controllers.portal.AuthoritativeSets.browse(id: String, params: services.search.SearchParams ?= services.search.SearchParams.empty, paging: utils.PageParams ?= utils.PageParams.empty)
 GET         /sets/:id/search                      @controllers.portal.AuthoritativeSets.search(id: String, params: services.search.SearchParams ?= services.search.SearchParams.empty, paging: utils.PageParams ?= utils.PageParams.empty, inline: Boolean ?= false)
